Change element ids to ints

+++ landlab/utils/structured_grid.py
@@ -676,7 +676,7 @@ def active_links(shape, node_status_array=None, link_nodes=None):
 
     (active_links, ) = np.where(active_links)
 
-    return active_links
+    return active_links.astype(np.int)
 
 
 def active_face_index(shape):
